move indirect hashing into its own package
