+++ src/accel/data/mod.rs
@@ -15,7 +15,6 @@ use arrow::{
 use pyo3::{exceptions::PyRuntimeError, prelude::*};
 
 mod index;
-mod indirect_hash;
 mod rc_set;
 mod selection;
 mod sorting;
